+++ nakadion/src/consumer/error.rs
@@ -165,7 +165,7 @@ impl ConsumerError {
     }
 
     pub fn message(&self) -> Option<&str> {
-        self.message.as_ref().map(|m| &**m)
+        self.message.as_deref()
     }
 
     pub(crate) fn enriched(self, batches_processed: usize) -> crate::internals::EnrichedErr {